Camera: Wait for the status tracker idle fence
In case, the status tracker idle fence gets released
while waiting on 'mPendingChangeSignal' during device
disconnect, then there is not going to be a way to
signal the potential state transition. The wait will
eventually timeout and that will increase the device
disconnect latency.
To reduce some of the potential latency, identify whether
we have a valid idle fence and if necessary wait for the
signal instead.

Bug: 246214490
Test: Camera CTS,
Partner verified
Google: 2719522
Change-Id: Iaac39291c8b969d3ca563b188c18a050a71e54ba

Signed-off-by: S I P U N <sipunkumar85@gmail.com>